Add unit tests for Cryostat external cryo geometry calculations
